ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.6.0
python-dotenv==1.0.0
httpx==0.25.2
openai==1.3.7
anthropic==0.7.8
redis==5.0.1
pymongo==4.6.0
python-multipart==0.0.6
jinja2==3.1.2
markdown==3.5.1
pygments==2.17.2 